Initial commit
There is a fair bit of learning in here.

lingui brings it's own babel with it, which automatically grabs any
available .babelrc. If there is anything in there it will break.

The workaround for this is doing a build command with --no-babelrc and
pass in the plugins you need for the build. See scripts.build in
package.json.

Between babel ignoring everything (like the locale folder) outside src
and Lingui picking up the translations in the dist folder, I needed to
move the locale folder into src and ignore the dist folder with the
localeDir and srcPathDirs options to the lingui config to package.json.

GraphQL declares the types and schema ahead of time, adding static
strings in the descriptions. Changing the descriptions seems to require
creating the schema and types per request. I'm not sure what the
performance cost of that is.

I'd be interested to know how that would get handled with Apollo server.